requirements.txt
numpy==1.23.5
pandas==1.5.2
pillow==7.2.0